Rename crate to gaea
In Greek mythology, Gaia (/ˈɡeɪ.ə/ or /ˈɡaɪ.ə/; from Ancient Greek Γαῖα,
a poetical form of Γῆ Gē, "land" or "earth"), also spelled Gaea
(/ˈdʒiːə/), is the personification of the Earth and one of the Greek
primordial deities. Gaia is the ancestral mother of all life: the primal
Mother Earth goddess.

The crate will be the low-level foundation of the Heph runtime, and
perhaps others.

Closes #98.

And yes, ``Gaia'' was taken.